Update get queries

+++ MOBoard.Common/Contractors/V1/Issue/IssueDto.cs
@@ -15,7 +15,7 @@ namespace MOBoard.Common.Contractors.V1.Issue
         public int IssueNumber { get; set; }
         public string IssueFullNumber { get; set; }
         public IEnumerable<IssueCommentDto> IssueComments { get; set; }
-        public int LoggedTime { get; set; }
+        public decimal LoggedTime { get; set; }
         public IssuePriorityLevel Priority { get; set; }
     }
